requirements.txt
# GUI Framework
customtkinter==5.2.0

# Document Processing
python-docx==0.8.11
python-pptx==0.6.21

# PDF Processing  
pypdf==3.17.0
pdf2docx==0.5.6

# Image Processing
Pillow==10.0.0
img2pdf==0.4.4

# Windows Office Integration (Optional - Windows only)
pywin32==306

# Enhanced PDF features
reportlab==4.0.4

# Optional: For drag and drop functionality
tkinterdnd2==0.3.0

# Additional utilities
pathlib2